html/syntax/writing/character-references/escapes-003.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="UTF-8"/>
<title>HTML5, character references: lowercase entity</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/html5/syntax.html#character-references'>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<link rel="stylesheet" type="text/css" href="../../../../resources/testharness.css">
<meta name='flags' content='dom'>
<meta name="assert" content="A lower case character entity reference produces the intended character.">
<style>
.test { color: white; }
</style>
</head>
<body>
<div class="test" id="test"><span id="one">&aacute;</span><span id="two" class="reference">รก</span></div>
<script>
test(function() {
assert_equals(document.getElementById("one").firstChild.data, document.getElementById("two").firstChild.data);
}, " ");
</script>
<div id='log'></div>
</body>
</html>
